requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
WTForms==2.3.3
Werkzeug==2.0.3
python-dotenv==0.19.2
SQLAlchemy==1.4.27
email_validator
PyJWT==2.1.0
Flask-JWT-Extended
email-validator
python-dotenv
flask_script
flask_limiter